#1-refactor: 훅 코드 분리
